crash--

+++ spiff/sensors/models.py
@@ -110,7 +110,9 @@ def exec_sensor_actions(sender, instance, created, **kwargs):
 post_save.connect(exec_sensor_actions, sender=SensorValue)
 
 def flush_old_values(sender, instance, created, **kwargs):
-  for v in instance.sensor.values.extra(order_by='stamp')[0:-instance.sensor.ttl]:
-    v.delete()
+  oldValues = instance.sensor.values.extra(order_by='stamp')
+  if len(oldValues) > instance.sensor.ttl:
+    for v in oldValues[0:len(oldValues)-instance.sensor.ttl]:
+      v.delete()
 
 post_save.connect(flush_old_values, sender=SensorValue)
